Refactor signal logic into its own module
