+++ packages/integrations/preact/package.json
@@ -35,8 +35,8 @@
     "@babel/core": ">=7.0.0-0 <8.0.0",
     "@babel/plugin-transform-react-jsx": "^7.17.12",
     "babel-plugin-module-resolver": "^4.1.0",
-    "preact-render-to-string": "^5.2.0",
-    "@preact/signals": "1.0.3"
+    "preact-render-to-string": "^5.2.4",
+    "@preact/signals": "^1.1.0"
   },
   "devDependencies": {
     "astro": "workspace:*",
