+++ packages/integrations/preact/src/context.ts
@@ -1,9 +1,10 @@
-import type { RendererContext, SignalLike } from './types';
+import type { RendererContext, SignalLike, PropNameToSignalMap } from './types';
 
-type Context = {
+export type Context = {
 	id: string;
 	c: number;
 	signals: Map<SignalLike, string>;
+	propsToSignals: Map<Record<string, any>, PropNameToSignalMap>;
 };
 
 const contexts = new WeakMap<RendererContext['result'], Context>();
@@ -17,7 +18,8 @@ export function getContext(result: RendererContext['result']): Context {
 		get id() {
 			return 'p' + this.c.toString();
 		},
-		signals: new Map()
+		signals: new Map(),
+		propsToSignals: new Map()
 	};
 	contexts.set(result, ctx);
 	return ctx;
