+++ packages/integrations/preact/src/types.ts
@@ -5,4 +5,10 @@ export type RendererContext = {
 
 export type SignalLike = {
 	peek(): any;
-}
+};
+
+export type PropNameToSignalMap = Map<string, SignalLike>;
+
+export type AstroPreactAttrs = {
+	['data-preact-signals']?: string
+};
